feat(KP-355): Search Filters are now combined
